refactor: packages rearrangement
