+++ src/main/java/io/naoki/currencyspring/dto/currency/CurrencyPair.java
@@ -1,4 +1,4 @@
-package io.naoki.currencyspring.currency;
+package io.naoki.currencyspring.dto.currency;
 
 public record CurrencyPair(
         String baseCurrencyCode,
