+++ src/main/java/io/naoki/currencyspring/service/impl/CurrencyNotFoundException.java
@@ -1,4 +1,4 @@
-package io.naoki.currencyspring.currency;
+package io.naoki.currencyspring.service.impl;
 
 public class CurrencyNotFoundException extends RuntimeException {
 }
